+++ src/interpreter/execute.cpp
@@ -3,6 +3,7 @@
 #include "../ast.hpp"
 #include "../ast_allocator.hpp"
 #include "../compute_offsets.hpp"
+#include "../convert_ast.hpp"
 #include "../cst_allocator.hpp"
 #include "../frontend_context.hpp"
 #include "../lexer.hpp"
